Correctly block reconnection to a websocket when revoking authorization (#289)
The previously comitted logic did correctly disconnect the user from the websocket, but missed blocking re-connecting with the same token. This updated logic correctly tracks the time of revocation and prevents any JWT issued to that user for that server before the revocation time from being re-used.

This was simpler than having to update the Panel to send along a bunch of JTIs and feels less error prone. I want to clean this up a lot, but it'll be a breaking change so we'll do it as a 2.0 release when we can safely do all of that.